main/requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
asyncer==0.0.2
asynctest==0.13.0
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
cycler==0.11.0
einops==0.6.1
exceptiongroup==1.2.0
fastapi==0.103.2
ffmpy==0.3.1
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gast==0.4.0
google-auth==2.25.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==3.34.0
gradio_client==0.2.6
grpcio==1.60.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.6
ImageHash==4.3.1
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.5
libclang==16.0.6
linkify-it-py==2.0.2
llvmlite==0.39.1
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.9.0.1
mpmath==1.3.0
multidict==6.0.4
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
oauthlib==3.2.2
onnxruntime==1.14.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.7
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
pooch==1.8.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pyheif==0.7.1
PyMatting==1.1.12
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==6.0.1
qudida==0.0.4
rembg==2.0.40
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.10.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2021.11.2
toolz==0.12.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.66.1
typing_extensions==4.7.1
uc-micro-py==1.0.2
urllib3==2.0.7
uvicorn==0.22.0
watchdog==3.0.0
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.15.0
